split RdpSettings
