+++ core/CMakeLists.txt
@@ -8,7 +8,9 @@ set(MSRDPEX_CORE_HEADERS
     "${MSRDPEX_INCLUDE_DIR}/OutputMirror.h"
     "${MSRDPEX_INCLUDE_DIR}/Process.h"
     "${MSRDPEX_INCLUDE_DIR}/RdpFile.h"
+    "${MSRDPEX_INCLUDE_DIR}/RdpInstance.h"
     "${MSRDPEX_INCLUDE_DIR}/RdpSession.h"
+    "${MSRDPEX_INCLUDE_DIR}/RdpSettings.h"
     "${MSRDPEX_INCLUDE_DIR}/VideoRecorder.h"
     "${MSRDPEX_INCLUDE_DIR}/MsRdpEx.h")
