+++ dll/RdpInstance.cpp
@@ -1,0 +1,81 @@
+
+#include <MsRdpEx/RdpInstance.h>
+
+extern "C" const GUID IID_IMsRdpExInstance;
+
+class CMsRdpExInstance : public IMsRdpExInstance
+{
+public:
+    CMsRdpExInstance(CMsRdpClient* pMsRdpClient)
+    {
+        m_refCount = 0;
+        m_pMsRdpClient = pMsRdpClient;
+    }
+
+    ~CMsRdpExInstance()
+    {
+
+    }
+
+    // IUnknown interface
+public:
+    HRESULT STDMETHODCALLTYPE QueryInterface(
+        REFIID riid,
+        LPVOID* ppvObject
+    )
+    {
+        HRESULT hr = E_NOINTERFACE;
+        MsRdpEx_Log("CMsRdpExInstance::QueryInterface");
+
+        if (riid == IID_IUnknown)
+        {
+            *ppvObject = (LPVOID)((IUnknown*)this);
+            m_refCount++;
+            return S_OK;
+        }
+        else if (riid == IID_IMsRdpExInstance)
+        {
+            *ppvObject = (LPVOID)((IUnknown*)this);
+            m_refCount++;
+            return S_OK;
+        }
+
+        return hr;
+    }
+
+    ULONG STDMETHODCALLTYPE AddRef()
+    {
+        MsRdpEx_Log("CMsRdpExInstance::AddRef");
+        return ++m_refCount;
+    }
+
+    ULONG STDMETHODCALLTYPE Release()
+    {
+        MsRdpEx_Log("CMsRdpExInstance::Release");
+        if (--m_refCount == 0)
+        {
+            MsRdpEx_Log("--> deleting object");
+            delete this;
+            return 0;
+        }
+        MsRdpEx_Log("--> refCount=%d", m_refCount);
+        return m_refCount;
+    }
+
+    // IMsRdpExInstance
+public:
+    HRESULT STDMETHODCALLTYPE GetRdpClient(LPVOID* ppvObject)
+    {
+        IUnknown* pMsRdpClient = (IUnknown*)m_pMsRdpClient;
+        return pMsRdpClient->QueryInterface(IID_IUnknown, ppvObject);
+    }
+
+private:
+    ULONG m_refCount;
+    CMsRdpClient* m_pMsRdpClient;
+};
+
+CMsRdpExInstance* CMsRdpExInstance_New(CMsRdpClient* pMsRdpClient)
+{
+    return new CMsRdpExInstance(pMsRdpClient);
+}
